WEEK6/requirements.txt
# Django project requirements
# ===========================

Django>=4.2
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
mysqlclient>=2.2
python-decouple>=3.8
bcrypt>=4.0
django-cors-headers>=4.3
